requirements.txt
dotenv
pyttsx3
PyAudio
selenium
playsound
wikipedia
wolframalpha
opencv-python
SpeechRecognition
google-generativeai
mysql-connector-python
sentence-transformers
faiss-cpu
PyPDF2
python-docx
python-pptx
openpyxl
pyperclip
pygetwindow